[dart2wasm] Make implementation libraries part of core to hide internals
Previously we had started to move core patch files to libraries for
better code organization.

However moving implementation classes to separate libraries required
making the implementation types and internal members public.

This has two problems:

1. The members can be accessed by end users via `dynamic`.
2. More importantly, because those public internal members can impact
   TFA results.

We don't care about (1) too much, but (2) is important.

So in this CL we make the libraries part of `core` again, while keeping
the file structure the same when possible.

With implementation files listed as "patches" they can still have their
own imports, but hey won't be importing each other as they're all part
of the same library.

This is the first CL in series that merges `_boxed_int` and
`_boxed_double` libraries back to `core`, and `_typed_data` back to
`typed_data`. Follow-up CLs will merge the other libraries back to
`core`.

The conventions for files is:

- Files that patch core types are named `..._patch.dart`.

- Implementation classes are in separate files. Names of the
  implementation files is either the implementation class name (e.g.
  `boxed_double.dart` for `_BoxedDouble`), or when implementing multiple
  class, name of the library being implemented (e.g. `typed_data.dart`).

Following these conventions, `double.dart` is renamed to
`double_patch.dart`.

Because we lose namespacing for helpers, the helper intrinsics like
`double` `_toInt` are moved to class members, as before.

One problem in the current CL is that in `libraries.yaml`, when I
`include` a target, I can't add more `core` patches to the included
target, the patch section overrides the included patches. I don't know
if this is a bug yet. If it is, we can refactor `libraries.yaml` after
it's fixed.

Tested: Changes covered by existing tests.
Change-Id: Iba7d81d383deff61aad521000f0ca9cf9276dcfb
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/319500
Reviewed-by: Aske Simon Christensen <askesc@google.com>
Reviewed-by: Joshua Litt <joshualitt@google.com>
Commit-Queue: Ömer Ağacan <omersa@google.com>