[vm/ffi] Consistently treat handles as int32 in IL on 32 bit archs
We were treating the handle inconsistently as int32 and uint32 on 32
bit platforms, leading to intconvertor instructions being inserted.

The marshaller was already documenting we don't ever expose the int
value of the handle to users, so there is no need to ever treat it as
uint32.

TEST=ffi/regress_jump_to_frame_test
TEST=ffi/vmspecific_handle_test

Closes: https://github.com/dart-lang/sdk/issues/53224

Change-Id: Ibf8a7922b8bf65c22767b0eb3994ba63643ce539
Cq-Include-Trybots: luci.dart.try:vm-linux-debug-ia32-try,vm-ffi-android-debug-arm-try
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/321081
Commit-Queue: Daco Harkes <dacoharkes@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>