[cfe] Avoid repearing type inference on pattern for-in loops
Previously the type inference was repeated in order to obtain the
static type of the iterable expression. That type is computed in the
shared analysis, and this CL passes the computed type to the clients
of the analysis routine in the result object.

In addition to being redundant, the repeated inference pass caused
crashes.

Closes https://github.com/dart-lang/sdk/issues/53092

Change-Id: I63d3d82be046884fa16d904a43e3bb6e01cf6f7e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/321141
Reviewed-by: Johnni Winther <johnniwinther@google.com>
Commit-Queue: Chloe Stefantsova <cstefantsova@google.com>