[analysis_server] Fix filterText for overrides
I noticed filterText was incorrect in some cases - for example if label was "MyClass(a)" filterText should be only "MyClass" without the "a".

There was a test for this but it was wrong, it checked filterText was null, even though label contained parens (a null filterText means use label).

I renamed the tests so they moved around a little, but they should be clearer (and more correct) now.

Change-Id: If94f80e8d7dc8b394c49eebb4d7482de2d62c38b
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/322461
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>