[linter] Add NON_BOOL_CONDITION to hasConstError.
Context:
`prefer_const_constructor` uses the constant evaluator to check if a
constructor can be made constant. There's a long standing issue with
`kIsWeb` that hasn't been fixed where `kIsWeb` is considered an
unknown bool, but the evaluator has a small exception to handle it
differently. With the recent changes, I suspect that we surface
NON_BOOL_CONDITION due to the way we handle `kIsWeb`, but it's not realized as a const error and we then dart fix these constructors which is wrong.

This CL adds NON_BOOL_CONDITION to the whitelist for the linter
and prevents producing `prefer_const_constructor` when it won't
make sense to.

Bug: https://github.com/flutter/flutter/issues/133152
Change-Id: Id5d47aec12a7fec741206d4785a9e2ba9a40b879
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/322785
Reviewed-by: Samuel Rawlins <srawlins@google.com>
Commit-Queue: Kallen Tu <kallentu@google.com>