Flow analysis: use virtual dispatch to access property target SSA nodes.
Previously, we used a switch statement in
`_FlowAnalysisImpl._handleProperty` with pattern matching. In a
follow-up CL, I will be shifting this logic to the callers of
`_handleProperty`. Implementing it as a virtual method dispatch rather
than a switch statement will mean that less code will need to be
duplicated.

There is a small performance improvement in CFE compilations:

instructions:u: -0.0192% +/- 0.0010% (-3708144.33 +/- 200427.86)
branches:u: -0.0332% +/- 0.0013% (-1309699.00 +/- 51985.25)

Change-Id: I3767917035ff0c462aa53f8023e72ce061211f2e
Bug: https://dart-review.googlesource.com/c/sdk/+/320701
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/320701
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>