[vm/compiler] Avoid searching static functions in the imported libraries
References to members are fully resolved in kernel. So, when
looking for a static function in the library scope, there is no need to
search in the imported libraries.

Searching in the imported libraries could be very slow as they can be
huge (e.g. the whole Flutter framework).

Fixes https://github.com/flutter/flutter/issues/133195
TEST=manually verified repro from https://github.com/flutter/flutter/issues/133195

Change-Id: Ia7e19a6ceaac291192c34b96587dad5cb4b80aad
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/323201
Reviewed-by: Ryan Macnak <rmacnak@google.com>
Commit-Queue: Alexander Markov <alexmarkov@google.com>