[dart2wasm] Avoid updating class ID field to allow it to be immutable
Change the intrinsics for the internal `makeListFixedLength` and
`makeFixedListUnmodifiable` functions to allocate a new object
(pointing to the existing array) instead of updating the class ID of
the existing object.

In combination with the the `struct.new` optimization, this enables
the class ID field to become immutable.

Change-Id: I981a83a3f89f57f6afa0ac9d952d2d662ab42fbd
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/323360
Reviewed-by: Jess Lally <jessicalally@google.com>
Commit-Queue: Aske Simon Christensen <askesc@google.com>