[dart2js] Fix regression in Dart2JS phase3 memory usage.
Due to a recent change to improve the dump info task, the `runSequentialPhases` method was reorganized. Part of this involved adding references to the closed world and global inference indices later on in the method (after the emitter task runs) to support serialization. Though this serialization only occurs if certain flags are passed to the compiler, the VM can no longer GC the results of those earlier phases as early since they might be used.

The fix is to reassign those locals earlier on based on the dump info flags. This way the VM will track the new locals separately and can free the original references while only adding a new reference if the relevant flags are passed.

Change-Id: I8f5dbc0f88c77ab43cb3f3bf12a42ecb37c9a803
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/324660
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Commit-Queue: Nate Biggs <natebiggs@google.com>